Minor edits to clarify terminology (/issues/399)
Line 69: Change "variable" to "data frame" for consistent terminology. Also change the word "print" as it will be confusing to those new to R (or coding in general) as it has connotations of sending stuff to a printer.
Line 109: Change "Because columns are vectors, they all contain the same class of data", because it sounds like all columns in a data frame contain the same class of data.
Line 313: Change the US colloquialism "under the hood".
Line 316: De-italicize (is that a word?!) "levels" the second time it was done.
Line 484: Insert "function" in place of "call" which is a fluid representation that hasn't be used elsewhere.