Use patchwork instead of gridExtra to arrange plots, improve example
patchwork has a much simpler syntax and is a tidyverse-adjacent package.
I also added installation instructions for the package, fixed a code error  in the example (`labs()` was added twice to the 1st plot)  and reduced the complexity of the syntax to better highlight the important parts (shorter variable name, fewer linebreaks)
I also gave "exporting plots" a separate headline, because it really is a separate section.